requirements.txt
# Installs bike_image_classifier_tensorflow as a local package. This makes it accessible
# via `import bike_image_classifier_tensorflow` in Python.
-e .

# Adds the whisk cli, directory helpers, etc.
whisk==0.1.26

### whisk project requirements
ipykernel # creating the `bike_image_classifier_tensorflow` kernel
python-dotenv # access env vars defined in a .env file from Python code
dvc # Data Version Control (DVC) for versioning data and pipelines
boto3 # AWS API for S3 bucket management needed for DVC remotes
Flask==1.1.2 # Web service HTTP API
honcho==1.0.1 # Procfile support
gunicorn==20.0.4 # App server for Flask app
tox==3.15.0 # Testing the package in insolation and across Python versions
twine==1.14.0 # Package distribution
pytest==4.6.5 # Tests
pytest-runner==5.1 # Tests

### Common Data Science Packages
# Already have a list of project requirements? Remove conflicting packages.
cloudpickle # For serializing constructs that Pickle can't.
ipython
jupyter
matplotlib
numpy
pandas
plotly
requests
seaborn
scikit-learn
scipy
simplejson
sqlalchemy

### Add project-specific requirements below. ###
absl-py==0.9.0
astunparse==1.6.3
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
gast==0.3.3
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
imutils==0.5.3
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
oauthlib==3.1.0
opencv-python-headless==4.2.0.34
opt-einsum==3.2.1
Pillow==7.1.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow-cpu==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
urllib3==1.25.9
validators==0.15.0
Werkzeug==1.0.1
wrapt==1.12.1
